backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
-e file:../cavlib
    # via -r requirements.in
appnope==0.1.0
    # via ipython
asgiref==3.4.1
    # via django
backcall==0.1.0
    # via ipython
cached-property==1.5.1
    # via -r requirements.in
click==7.0
    # via pip-tools
cycler==0.10.0
    # via matplotlib
decorator==4.4.1
    # via
    #   ipython
    #   networkx
    #   traitlets
django==3.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.2.1
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.11.0
    # via django-rest-framework
entrypoints==0.3
    # via flake8
flake8==3.7.9
    # via -r requirements.in
gunicorn[gthread]==20.0.4
    # via -r requirements.in
imageio==2.8.0
    # via scikit-image
ipython==7.12.0
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
jedi==0.16.0
    # via ipython
joblib==0.14.1
    # via scikit-learn
kiwisolver==1.1.0
    # via matplotlib
matplotlib==3.1.3
    # via scikit-image
mccabe==0.6.1
    # via flake8
methodtools==0.1.2
    # via -r requirements.in
msgpack==1.0.0
    # via
    #   -r requirements.in
    #   cavlib
networkx==2.4
    # via scikit-image
numpy==1.18.1
    # via
    #   -r requirements.in
    #   cavlib
    #   imageio
    #   matplotlib
    #   pywavelets
    #   scikit-learn
    #   scipy
    #   tflite-runtime
parso==0.6.2
    # via jedi
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==7.0.0
    # via
    #   -r requirements.in
    #   cavlib
    #   imageio
    #   scikit-image
pip-tools==6.6.2
    # via -r requirements.in
platformdirs==2.4.0
    # via -r requirements.in
prompt-toolkit==3.0.3
    # via ipython
ptyprocess==0.6.0
    # via pexpect
pycodestyle==2.5.0
    # via flake8
pyflakes==2.1.1
    # via flake8
pygments==2.5.2
    # via ipython
pyinstrument==3.1.3
    # via -r requirements.in
pyinstrument-cext==0.2.2
    # via pyinstrument
pyparsing==2.4.6
    # via matplotlib
python-dateutil==2.8.1
    # via matplotlib
pytz==2019.3
    # via django
pywavelets==1.1.1
    # via scikit-image
scikit-image==0.16.2
    # via
    #   -r requirements.in
    #   cavlib
scikit-learn==0.22.2
    # via
    #   -r requirements.in
    #   cavlib
scipy==1.4.1
    # via
    #   scikit-image
    #   scikit-learn
six==1.14.0
    # via
    #   cycler
    #   python-dateutil
    #   traitlets
    #   wirerope
sqlparse==0.3.0
    # via django
tflite-runtime==2.5.0
    # via
    #   -r requirements.in
    #   cavlib
tomli==1.2.1
    # via pep517
tqdm==4.43.0
    # via -r requirements.in
traitlets==4.3.3
    # via ipython
typing-extensions==4.2.0
    # via
    #   -r requirements.in
    #   cavlib
wcwidth==0.1.8
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
wirerope==0.3.1
    # via methodtools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
